Hide a framework7 navbar

Hide a framework7 navbar

f7NavbarHide(
  session = shiny::getDefaultReactiveDomain(),
  animate = TRUE,
  hideStatusbar = FALSE
)

Arguments

session
Shiny session object.

animate
Whether it should be hidden with animation or not. By default is TRUE.

hideStatusbar
When FALSE (default) it hides navbar partially keeping space required to cover statusbar area. Otherwise, navbar will be fully hidden.

Examples

if (interactive()) {
 library(shiny)
 library(shinyMobile)

 shiny::shinyApp(
    ui = f7Page(
       title = "Accordions",
       f7SingleLayout(
          navbar = f7Navbar("Hide/Show navbar"),
          f7Segment(
             f7Button(inputId = "hide", "Hide navbar", color = "red"),
             f7Button(inputId = "show", "Show navbar", color = "green"),
          )
       )
    ),
    server = function(input, output, session) {

       observeEvent(input$hide, {
          f7NavbarHide()
       })

       observeEvent(input$show, {
          f7NavbarShow()
       })
    }
 )
}