shinyMobile: Mobile Ready 'shiny' Apps with Standalone Capabilities

Develop outstanding 'shiny' apps for 'iOS' and 'Android' as well as beautiful 'shiny' gadgets. 'shinyMobile' is built on top of the latest 'Framework7' template https://framework7.io. Discover 14 new input widgets (sliders, vertical sliders, stepper, grouped action buttons, toggles, picker, smart select, ...), 2 themes (light and dark), 12 new widgets (expandable cards, badges, chips, timelines, gauges, progress bars, ...) combined with the power of server-side notifications such as alerts, modals, toasts, action sheets, sheets (and more) as well as 3 layouts (single, tabs and split).

See also

Author

Maintainer: David Granjon dgranjon@ymail.com

Authors:

Veerle van Leemput veerle@hypebright.nl

Victor Perrier info@dreamrs.fr

Isabelle Rudolf

Other contributors:

AthlyticZ [funder]

John Coene jcoenep@gmail.com [contributor]

Dieter Menne dieter.menne@menne-biomed.de [contributor]

Marvelapp (device.css wrappers) [contributor, copyright holder]

Vladimir Kharlampidi nolimits4web@gmail.com (Framework7 HTML template) [contributor, copyright holder]