Create a manifest for your shiny app

This is a central piece if you want to have your app standalone for instance

create_manifest(
  path,
  name = "My App",
  shortName = "My App",
  description = "What it does!",
  lang = "en-US",
  startUrl,
  display = c("minimal-ui", "standalone", "fullscreen", "browser"),
  icons
)

Arguments

path
package path.

name
App name.

shortName
App short name.

description
App description

lang
App language (en-US by default).

startUrl
Page to open at start.

display
Display mode. Choose among c("minimal-ui", "standalone", "fullscreen", "browser"). In practice, you want the standalone mode so that the app looks like a native app.

icons
Dataframe containing icons specs for instance data.frame(src = rep("icons/128x128.png", 10), sizes = rep("128x128", 10), types = rep("image/png", 10)). src gives the icon path (in the www folder for instance), sizes gives the size and types the type.

Value

This function creates a www folder for your shiny app. Must specify the path. It also creates 2 additional folders to contain icons and splashscreen (splashscreen creation is not automatic for iOS). It also creates the manifest.json file in this www folder. You still have to create your own icons and splashscreen!

Note

See https://developer.mozilla.org/en-US/docs/Web/Manifest for more informations.

Examples

create_manifest(
  path = tempdir(),
  name = "My App",
  shortName = "My App",
  description = "What it does!",
  lang = "en-US",
  startUrl = "https://www.google.com/",
  display = "standalone",
  icons = data.frame(
    src = rep("icons/128x128.png", 10),
    sizes = rep("128x128", 10),
    types = rep("image/png", 10)
  )
)

#> {
#>   "name": "My App",
#>   "short_name": "My App",
#>   "description": "What it does!",
#>   "lang": "en-US",
#>   "start_url": "https://www.google.com/",
#>   "display": "standalone",
#>   "icons": [
#>     {
#>       "src": "icons/128x128.png",
#>       "sizes": "128x128",
#>       "types": "image/png"
#>     },
#>     {
#>       "src": "icons/128x128.png",
#>       "sizes": "128x128",
#>       "types": "image/png"
#>     },
#>     {
#>       "src": "icons/128x128.png",
#>       "sizes": "128x128",
#>       "types": "image/png"
#>     },
#>     {
#>       "src": "icons/128x128.png",
#>       "sizes": "128x128",
#>       "types": "image/png"
#>     },
#>     {
#>       "src": "icons/128x128.png",
#>       "sizes": "128x128",
#>       "types": "image/png"
#>     },
#>     {
#>       "src": "icons/128x128.png",
#>       "sizes": "128x128",
#>       "types": "image/png"
#>     },
#>     {
#>       "src": "icons/128x128.png",
#>       "sizes": "128x128",
#>       "types": "image/png"
#>     },
#>     {
#>       "src": "icons/128x128.png",
#>       "sizes": "128x128",
#>       "types": "image/png"
#>     },
#>     {
#>       "src": "icons/128x128.png",
#>       "sizes": "128x128",
#>       "types": "image/png"
#>     },
#>     {
#>       "src": "icons/128x128.png",
#>       "sizes": "128x128",
#>       "types": "image/png"
#>     }
#>   ]
#> }