Create a manifest for your shiny app

Deprecated. Please use this workflow instead: https://unleash-shiny.rinterface.com/mobile-pwa.html#charpente-and-pwa-tools.

create_manifest(
  path,
  name = "My App",
  shortName = "My App",
  description = "What it does!",
  lang = "en-US",
  startUrl,
  display = c("minimal-ui", "standalone", "fullscreen", "browser"),
  background_color = "#000000",
  theme_color = "#0000ffff",
  icon
)

Arguments

path
package path.

name
App name.

shortName
App short name.

description
App description

lang
App language (en-US by default).

startUrl
Page to open at start.

display
Display mode. Choose among c("minimal-ui", "standalone", "fullscreen", "browser"). In practice, you want the standalone mode so that the app looks like a native app.

background_color
The background_color property is used on the splash screen when the application is first launched.

theme_color
The theme_color sets the color of the tool bar, and may be reflected in the app's preview in task switchers.

icon
Dataframe containing icon specs. src gives the icon path (in the www folder for instance), sizes gives the size and types the type.

Value

This function creates a www folder for your shiny app. Must specify the path. It creates 1 folders to contain icons and the manifest.json file.

Note

See https://developer.mozilla.org/en-US/docs/Web/Manifest for more informations.

Examples

create_manifest(
  path = tempdir(),
  name = "My App",
  shortName = "My App",
  description = "What it does!",
  lang = "en-US",
  startUrl = "https://www.google.com/",
  display = "standalone",
  background_color = "#3367D6",
  theme_color = "#3367D6",
  icon = data.frame(
    src = "icons/128x128.png",
    sizes = "128x128", 10,
    types = "image/png"
  )
)

#> Warning: create_manifest will be removed in future release. Please use
#>     the workflow desribed at https://unleash-shiny.rinterface.com/mobile-pwa.html#charpente-and-pwa-tools instead.

#> {
#>   "name": "My App",
#>   "short_name": "My App",
#>   "description": "What it does!",
#>   "lang": "en-US",
#>   "start_url": "https://www.google.com/",
#>   "display": "standalone",
#>   "background_color": "#3367D6",
#>   "theme_color": "#3367D6",
#>   "icon": [
#>     {
#>       "src": "icons/128x128.png",
#>       "sizes": "128x128",
#>       "X10": 10,
#>       "types": "image/png"
#>     }
#>   ]
#> }