Framework7 Toolbar

f7Toolbar is a layout element located at the bottom or top. It is internally used by f7Tabs.

f7Toolbar(
  ...,
  position = c("top", "bottom"),
  hairline = TRUE,
  shadow = TRUE,
  icons = FALSE,
  scrollable = FALSE
)

Arguments

...
Slot for f7Link or any other element.

position
Tabs position: "top" or "bottom".

hairline
Whether to display a thin border on the top of the toolbar. TRUE by default.

shadow
Whether to display a shadow. TRUE by default.

icons
Whether to use icons instead of text. Either ios or md icons.

scrollable
Whether to allow scrolling. FALSE by default.

Author

David Granjon, dgranjon@ymail.com